refactor : IDE-117 파일 생성 함수 커스텀 hook 분리
