+++ src/app/ide/FileTree/Node.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 import axiosInstance from '@/app/api/axiosInstance';
 import { useFileTreeStore } from '@/store/useFileTreeStore';
 import { FileNodeType } from '@/types/IDE/FileTree/FileDataTypes';
-import useCurrentOpenFileList from '@/store/useCurrentOpenFile';
+
 import {
   findLanguage,
   findNowFilePath,
@@ -17,6 +17,7 @@ import {
 
 import useHandleOpenFile from '@/app/hooks/useHandleOpenFile';
 import LanguageIcon from './LanguageIcon';
+import useHandleCreateFile from '@/app/hooks/useHandleCreateFile';
 
 export const Node = ({
   node,
@@ -26,45 +27,16 @@ export const Node = ({
 }: NodeRendererProps<FileNodeType>) => {
   const handleOpenFile = useHandleOpenFile();
   const { updateNodeName } = useFileTreeStore();
-  const { setOpenFilesIdList } = useCurrentOpenFileList();
+  const handleCreateFileRequest = useHandleCreateFile(node);
+
   const projectId = 'ebc63279-89b9-4b1d-bb4d-1270130c3d4d'; //임시
 
   const onNodeClick = (node: NodeApi<FileNodeType>) => {
     handleOpenFile(node);
   };
-  const handleCreateFileRequest = async (newNodeName: string) => {
-    try {
-      const nowFilePath = findNowFilePath(node) + newNodeName;
-      let responseData;
 
-      //directory or file 구별
-      if (node.isInternal) {
-        const response = await axiosInstance.post('/api/files', {
-          projectId: projectId,
-          directories: nowFilePath,
-          files: null,
-          content: 'print("Hello, World!")',
-        });
-        responseData = response.data;
-      } else {
-        const response = await axiosInstance.post('/api/files', {
-          projectId: projectId,
-          directories: null,
-          files: nowFilePath,
-          content: 'print("Hello, World!")',
-        });
-        responseData = response.data;
-      }
-
-      //응답받은 filename, content .. 등 필요 정보 담아두기 -> 총미
-
-      //열린 파일 목록 업데이트 -> 총미
-      setOpenFilesIdList(node.id);
-
-      return responseData;
-    } catch (error) {
-      console.error(error);
-    }
+  const onCreateFile = async (newNodeName: string) => {
+    await handleCreateFileRequest(newNodeName);
   };
 
   const handleDeleteFileRequest = async () => {
@@ -143,7 +115,7 @@ export const Node = ({
                 if (e.key === 'Escape') node.reset();
                 if (e.key === 'Enter') {
                   if (isCorrectName(e.currentTarget.value) === true) {
-                    handleCreateFileRequest(e.currentTarget.value);
+                    onCreateFile(e.currentTarget.value);
                     updateNodeName(node.id, e.currentTarget.value);
                     node.submit(e.currentTarget.value); //이때 서버로도 메시지 보내야 함
                   } else {
